Remove 'integration-tests' profile and explicitly name what the mocked database configuration is needed for
